Move hases test to its own file
